+++ CHANGELOG.md
@@ -3,6 +3,9 @@
     Placeholder for the next version (at the beginning of the line):
     ## **WORK IN PROGRESS**
 -->
+## **WORK IN PROGRESS**
+* `iobroker`: Add flag to prevent prerelease versions in `io-package.json`
+
 ## 3.1.1 (2021-09-16)
 * Execute `manual-review` plugin at the start of the commit phase instead of at the end of the edit phase.